+++ tests/testthat/test-spread2.R
@@ -1,7 +1,7 @@
 test_that("spread2 tests", {
 
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
   data.table::setDTthreads(1)
 
   # inputs for x
@@ -387,7 +387,7 @@ test_that("spread2 tests -- asymmetry", {
 
   withr::local_package("terra")
   withr::local_package("CircStats")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
 
   aOrig <- terra::rast(terra::ext(0, 100, 0, 100), res = 1)
 
@@ -615,7 +615,7 @@ test_that("spread2 tests -- asymmetry", {
 
 test_that("spread2 returnFrom", {
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
 
   aOrig <- terra::rast(terra::ext(0, 100, 0, 100), res = 1)
 
@@ -645,7 +645,7 @@ test_that("spread2 tests", {
   skip_if_not_installed("CircStats")
 
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
 
   aOrig <- terra::rast(terra::ext(0, 100, 0, 100), res = 1)
 
@@ -675,7 +675,7 @@ test_that("spread2 tests", {
 
 test_that("spread2 works with terra", {
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
 
   aOrig <- terra::rast(terra::ext(0, 100, 0, 100), res = 1)
 
@@ -702,7 +702,7 @@ test_that("spread2 works with terra", {
 
 test_that("spread2 tests -- persistence", {
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
 
   aOrig <- terra::rast(terra::ext(0, 50, 0, 50), res = 1)
 
@@ -744,7 +744,7 @@ test_that("spread2 tests -- persistence", {
 
 test_that("spread2 tests -- SpaDES.tools issue #22 NA in spreadProb", {
   withr::local_package("terra")
-  rastDF <- needTerraAndRaster() #
+  rastDF <- needTerraAndRaster()
 
   aOrig <- terra::rast(terra::ext(0, 50, 0, 50), res = 1)